requirements.txt

awesome-slugify
django-ordered-model
